# src/main/resources/assets/api-docs.yaml
swagger: '2.0'
info:
  title: Consent
  description: A set of web-services containing methods for ingesting and retrieving expressions representing data use restrictions found within consent forms.
  version: 1.0.1
  license:
    name: BSD
    url: 'http://opensource.org/licenses/BSD-3-Clause'
  termsOfService: 'https://github.com/DataBiosphere/consent'
securityDefinitions:
  consent_auth:
    type: oauth2
    authorizationUrl: 'https://accounts.google.com/o/oauth2/auth'
    flow: implicit
    scopes:
      profile: profile
      email: email
paths:
...
  /api/dac/{dacId}:
    get:
      summary: Get DAC by ID
      description: Get Data Access Committee by Id
      parameters:
        - name: dacId
          in: path
          description: The DAC ID
          required: true
          type: integer
      tags:
        - DAC
      responses:
        200:
          description: The Data Access Committee matching the provided ID
          schema:
            $ref: '#/definitions/Dac'
        404:
          description: No DAC found with the given ID
        500:
          description: Internal Server Error
    delete:
      summary: Delete DAC by ID
      description: Delete Data Access Committee by Id
      parameters:
        - name: dacId
          in: path
          description: The DAC ID
          required: true
          type: integer
      tags:
        - DAC
      responses:
        200:
          description: The Data Access Committee matching the provided ID was deleted
          schema:
            $ref: '#/definitions/Dac'
        404:
          description: No DAC found with the given ID
        500:
          description: Internal Server Error
...
  /api/dacuser/{email}:
    get:
      summary: describe
      description: Returns the user identified by the email.
      parameters:
        - name: email
          in: path
          description: The email of the user.
          required: true
          type: string
      tags:
        - User
        - Roles
      responses:
        200:
          description: Returns the user.
          schema:
            $ref: '#/definitions/User'
    delete:
      summary: delete
      description: Deletes the user identified by the email.
      parameters:
        - name: email
          in: path
          description: The email of the user.
          required: true
          type: string
      tags:
        - User
        - Roles
      responses:
        200:
          description: Response = User was deleted
          schema:
            type: string
    put:
      summary: update
      description: Updates the user identified by the ID.
      parameters:
        - name: email
          in: path
          description: The email of the user.
          required: true
          type: string
        - name: dac
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      tags:
        - User
        - Roles
      responses:
        200:
          description: Returns the updated user.
          schema:
            $ref: '#/definitions/User'
        400:
          description: Malformed entity.
...
  /api/report/user/{userId}:
        get:
          summary: describeAllReportsByUser
          description: Returns a collection of the HelpReports made by the specified user.
          parameters:
            - name: userId
              in: path
              description: UserId
              required: true
              type: integer
              format: int32
          tags:
            - Report
          responses:
             200:
                description: An array of the HelpReports made by the user.
                schema:
                  type: array
                  items:
                    $ref: '#/definitions/HelpReport'
...
definitions:
...
  Dac:
    type: object
    properties:
      dacId:
        type: integer
        description: The DAC Id
      name:
        type: string
        description: The DAC Name
      description:
        type: string
        description: The DAC Description
      createDate:
        type: string
        format: date
        description: Date created
      updateDate:
        type: string
        format: date
        description: Date last updated
  User:
    type: object
    properties:
      dacUserId:
        type: integer
        description: ID of the user
      email:
        type: string
        description: Email of the user
      displayName:
        type: string
        description: Name of the user
      createDate:
        type: string
        format: date
        description: Describes the date the User was created.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
  UserRole:
    type: object
    properties:
      roleId:
        type: integer
        format: int32
        description: ID of the role
      name:
        type: string
        description: Name of the role
      emailPreference:
        type: boolean
        description: The user wants email notifications
      status:
        type: string
        description: Status of the role, approved or not.
      rationale:
        type: string
        description: Reason why the role was not approved.
      profileCompleted:
        type: boolean
        description: The profile is completed. Researchers only.
...
  HelpReport:
    type: object
    properties:
      reportId:
        type: integer
        format: int32
        description: ID of the report.
      userName:
        type: string
        description: Name of user that made the report.
      createDate:
        type: string
        format: date
        description: Describes the date the Report was created.
      subject:
         type: string
         description: Subject of the report.
      description:
         type: string
         description: Content of the report.
      userId:
         type: integer
         format: int32
         description: ID of user that made the report.
...